// common.types.ts
export interface ProjectForm {
    name: string;
    branchName: string;
    accountNumber: string;
    swiftCode: string;
    dateOfBirth: string;
    firstApplicantSignatureImage: string;
    firstName: string;
    governmentId: string;
    governmentIdImage: string;
    lastName: string;
    middleName: string;
    nextOfKeenImage: string;
    occupation: string;
    phoneNumber: string;
    residence: string;
    sourceOfFunds: string;
}

export interface BankInfoForm {
    name: string;
    branchName: string;
    accountNumber: string;
    swiftCode: string;
}